face-recognition/face_recognition/requirements.txt
pytz
pillow
requests
gunicorn
django>2.0
django-environ
whitenoise
psycopg2-binary
raven
django-statsd-mozilla
statsd
